+++ spec/types/hash_spec.rb
@@ -536,8 +536,8 @@ describe Attributor::Hash do
         proc do
           key 'integer', Integer, null: false # non-nullable, but not required
           key 'datetime', DateTime
-          key 'not-optional', String, required: true # Means: present: true, null: false
-          key 'required-but-nullable', String, present: true, null: true
+          key 'not-optional', String, required: true
+          key 'required-but-nullable', String, required: true, null: true
         end
       end
 
@@ -588,7 +588,6 @@ describe Attributor::Hash do
             key 'consistency', Attributor::Boolean
             key 'availability', Attributor::Boolean, null: true
             key 'partitioning', Attributor::Boolean, null: true
-            # required DSL it mimics adding present: true in the key (but does not touch the null: option)
             requires 'consistency', 'availability'
             requires.all 'name'
           end
@@ -819,7 +818,7 @@ describe Attributor::Hash do
       context 'merging requires.all with attribute required: true' do
         let(:block) do
           proc do
-            key 'required string', String, present: true
+            key 'required string', String, required: true
             key '1', Integer
             key 'some_date', DateTime
             requires do
@@ -827,7 +826,7 @@ describe Attributor::Hash do
             end
           end
         end
-        it 'includes attributes with present: true into the :all requirements' do
+        it 'includes attributes with required: true into the :all requirements' do
           req_all = description[:requirements].select { |r| r[:type] == :all }.first
           expect(req_all[:attributes]).to include('required string', 'some_date')
         end
@@ -836,11 +835,11 @@ describe Attributor::Hash do
       context 'creates the :all requirement when any attribute has required: true' do
         let(:block) do
           proc do
-            key 'required string', String, required: true # alias for present: true and null: false
-            key 'required integer', Integer, present: true
+            key 'required string', String, required: true
+            key 'required integer', Integer, required: true
           end
         end
-        it 'includes attributes with present: true into the :all requirements' do
+        it 'includes attributes with required: true into the :all requirements' do
           req_all = description[:requirements].select { |r| r[:type] == :all }.first
           expect(req_all).not_to be(nil)
           expect(req_all[:attributes]).to include('required string', 'required integer')
@@ -1290,7 +1289,7 @@ describe Attributor::Hash do
             end
           end
         end
-        it 'includes attributes with present: true into :required' do
+        it 'includes attributes with required: true into :required' do
           expect(description[:required].size).to eq(2)
           expect(description[:required]).to include( 'required string','some_date' )
         end